+++ components/modules/app-shell/status-bar/components/ElectronUpdateIndicator.stories.ts
@@ -47,6 +47,10 @@ const installElectronMock = () => {
       maximize: async () => undefined,
       close: async () => undefined,
       pickSqliteFile: async () => null,
+      pickSaveFile: async () => null,
+      pickDirectory: async () => null,
+      writeFile: async () => undefined,
+      openPath: async () => '',
       getStoragePath: async () => '',
       openStoragePath: async () => undefined,
       getLogPath: async () => '',
