+++ components/modules/database-tools/backup-restore/containers/BackupRestoreContainer.vue
@@ -1,19 +1,24 @@
 <script setup lang="ts">
 import LoadingOverlay from '~/components/base/loading-overlay/LoadingOverlay.vue';
+import { useWorkspaceConnectionRoute } from '~/core/composables/useWorkspaceConnectionRoute';
 import { getDatabaseClientLabel } from '~/core/constants/database-backup';
+import { DatabaseClientType } from '~/core/constants/database-client-type';
+import { isDesktopApp } from '~/core/helpers/environment';
 import { useSchemaStore } from '~/core/stores';
 import { useManagementConnectionStore } from '~/core/stores/managementConnectionStore';
-import type { ExportOptions } from '~/core/types';
+import type { ExportOptions, NativeBackupToolName } from '~/core/types';
 import BackupProfileCard from '../components/BackupProfileCard.vue';
 import BackupRestoreUnsupportedAlert from '../components/BackupRestoreUnsupportedAlert.vue';
 import ExportOptionsForm from '../components/ExportOptionsForm.vue';
 import ImportFileDropzone from '../components/ImportFileDropzone.vue';
 import ImportOptionsForm from '../components/ImportOptionsForm.vue';
+import NativeToolSelector from '../components/NativeToolSelector.vue';
 import RestoreConfirmDialog from '../components/RestoreConfirmDialog.vue';
 import TransferProgressCard from '../components/TransferProgressCard.vue';
 import { useDatabaseExport } from '../hooks/useDatabaseExport';
 import { useDatabaseImport } from '../hooks/useDatabaseImport';
 import { useNativeBackupCapability } from '../hooks/useNativeBackupCapability';
+import { useNativeBackupToolSelection } from '../hooks/useNativeBackupToolSelection';
 import type {
   BackupRestoreSection,
   BackupRestoreTab,
@@ -27,6 +32,7 @@ interface Props {
 const props = defineProps<Props>();
 
 // ─── Connection ────────────────────────────────────────────────────────────────
+const { workspaceId } = useWorkspaceConnectionRoute();
 const connectionStore = useManagementConnectionStore();
 const schemaStore = useSchemaStore();
 
@@ -51,6 +57,7 @@ const {
 } = useNativeBackupCapability(connectionType);
 
 const isRefreshing = ref(false);
+const desktopRuntimeSelectionEnabled = computed(() => isDesktopApp());
 
 const handleReload = async () => {
   isRefreshing.value = true;
@@ -61,13 +68,23 @@ const handleReload = async () => {
   }
 };
 const nativeBackupSupported = computed(
-  () => capability.value?.available ?? false
+  () => capability.value?.supported ?? false
 );
 const exportAvailable = computed(
   () => capability.value?.exportAvailable ?? false
 );
 const importAvailable = computed(
   () => capability.value?.importAvailable ?? false
+);
+const exportSectionEnabled = computed(
+  () =>
+    nativeBackupSupported.value &&
+    (exportAvailable.value || desktopRuntimeSelectionEnabled.value)
+);
+const importSectionEnabled = computed(
+  () =>
+    nativeBackupSupported.value &&
+    (importAvailable.value || desktopRuntimeSelectionEnabled.value)
 );
 const supportMessage = computed(() => {
   if (capabilityError.value) {
@@ -114,10 +131,61 @@ const partialAvailabilityMessage = computed(() => {
 });
 const availableSchemas = computed(() => {
   if (!props.connectionId) return [];
+  const wsId = workspaceId.value || connectionData.value?.workspaceId;
+
   return (schemaStore.schemas[props.connectionId] || [])
-    .filter(s => s.connectionId === props.connectionId)
+    .filter(
+      s =>
+        s.connectionId === props.connectionId &&
+        (!wsId || s.workspaceId === wsId)
+    )
     .map(s => s.name);
 });
+const isSchemaLoading = computed(
+  () => schemaStore.loading[props.connectionId] || false
+);
+const schemaLoadError = ref<string | null>(null);
+
+const loadAvailableSchemas = async () => {
+  const connection = connectionData.value;
+  const wsId = workspaceId.value || connection?.workspaceId;
+
+  if (!props.connectionId || !wsId || !connection) {
+    return;
+  }
+
+  const hasSchemasForContext = (
+    schemaStore.schemas[props.connectionId] || []
+  ).some(
+    schema =>
+      schema.connectionId === props.connectionId && schema.workspaceId === wsId
+  );
+
+  if (hasSchemasForContext) {
+    schemaLoadError.value = null;
+    return;
+  }
+
+  try {
+    schemaLoadError.value = null;
+    await schemaStore.fetchSchemas({
+      connectionId: props.connectionId,
+      workspaceId: wsId,
+      connection,
+    });
+  } catch (error) {
+    schemaLoadError.value =
+      error instanceof Error ? error.message : 'Failed to load schemas.';
+  }
+};
+
+watch(
+  [connectionData, () => workspaceId.value],
+  () => {
+    void loadAvailableSchemas();
+  },
+  { immediate: true }
+);
 
 // ─── Tab management ────────────────────────────────────────────────────────────
 const sections: BackupRestoreSection[] = [
@@ -148,7 +216,7 @@ watch(
 );
 
 watch(
-  [activeTab, exportAvailable, importAvailable],
+  [activeTab, exportSectionEnabled, importSectionEnabled],
   ([tab, canExport, canImport]) => {
     if (tab === 'export' && !canExport && canImport) {
       activeTab.value = 'import';
@@ -178,8 +246,45 @@ const {
   reset: resetExport,
 } = useDatabaseExport(connectionData, capability);
 
+const buildToolChoices = (available: string[] = [], missing: string[] = []) =>
+  [...new Set([...available, ...missing])] as NativeBackupToolName[];
+
+const exportToolChoices = computed(() =>
+  buildToolChoices(
+    capability.value?.availableExportTools || [],
+    capability.value?.missingExportTools || []
+  )
+);
+const exportDetectedToolPaths = computed(
+  () => capability.value?.availableExportToolPaths || []
+);
+const {
+  selection: exportToolSelection,
+  filteredToolPaths: exportToolPaths,
+  runtimeSelection: exportRuntimeSelection,
+  canSubmit: canExportWithRuntime,
+} = useNativeBackupToolSelection(
+  exportDetectedToolPaths,
+  exportToolChoices,
+  desktopRuntimeSelectionEnabled
+);
+const openExportSaveDirectory = async () => {
+  const directoryPath = lastExport.value?.saveDirectoryPath;
+
+  if (!directoryPath || !window.electronAPI?.window.openPath) {
+    return;
+  }
+
+  await window.electronAPI.window.openPath(directoryPath);
+};
+
 const onExport = async (options: ExportOptions) => {
-  await exportDatabase(databaseName.value, options);
+  await exportDatabase(databaseName.value, {
+    options,
+    runtime: exportRuntimeSelection.value,
+    promptForSaveLocation: desktopRuntimeSelectionEnabled.value,
+    saveDirectoryPath: undefined,
+  });
 };
 
 // ─── Import ────────────────────────────────────────────────────────────────────
@@ -201,6 +306,42 @@ const {
   reset: resetImport,
 } = useDatabaseImport(connectionData, capability);
 
+const baseImportToolChoices = computed(() =>
+  buildToolChoices(
+    capability.value?.availableImportTools || [],
+    capability.value?.missingImportTools || []
+  )
+);
+const importToolChoices = computed(() => {
+  if (
+    connectionType.value !== DatabaseClientType.POSTGRES ||
+    !selectedFile.value?.name
+  ) {
+    return baseImportToolChoices.value;
+  }
+
+  return [
+    selectedFile.value.name.toLowerCase().endsWith('.dump')
+      ? 'pg_restore'
+      : 'psql',
+  ] as NativeBackupToolName[];
+});
+const importDetectedToolPaths = computed(() =>
+  (capability.value?.availableImportToolPaths || []).filter(toolPath =>
+    importToolChoices.value.includes(toolPath.tool)
+  )
+);
+const {
+  selection: importToolSelection,
+  filteredToolPaths: importToolPaths,
+  runtimeSelection: importRuntimeSelection,
+  canSubmit: canImportWithRuntime,
+} = useNativeBackupToolSelection(
+  importDetectedToolPaths,
+  importToolChoices,
+  desktopRuntimeSelectionEnabled
+);
+
 // ─── Restore confirm dialog ──────────────────────────────────────────────────────────
 const showRestoreConfirm = ref(false);
 
@@ -210,7 +351,7 @@ const onRestoreClick = () => {
 
 const onRestoreConfirmed = async () => {
   showRestoreConfirm.value = false;
-  await importDatabase();
+  await importDatabase(importRuntimeSelection.value);
 };
 
 const onRestoreCancelled = () => {
@@ -250,7 +391,9 @@ const onRestoreCancelled = () => {
             class="min-w-fit shrink-0 cursor-pointer rounded-sm"
             :disabled="
               isCapabilityLoading ||
-              (section.id === 'export' ? !exportAvailable : !importAvailable)
+              (section.id === 'export'
+                ? !exportSectionEnabled
+                : !importSectionEnabled)
             "
             @click="selectToolTab(section.id)"
           >
@@ -325,18 +468,6 @@ const onRestoreCancelled = () => {
             </Button>
           </Alert>
 
-          <Alert v-if="lastExport && !isExporting && !exportError">
-            <Icon
-              name="hugeicons:checkmark-circle-02"
-              class="size-4 text-green-500"
-            />
-            <AlertTitle>Backup Complete</AlertTitle>
-            <AlertDescription>
-              {{ lastExport?.fileName }} downloaded in
-              {{ ((lastExport?.duration || 0) / 1000).toFixed(1) }}s
-            </AlertDescription>
-          </Alert>
-
           <TransferProgressCard
             v-if="isExporting || exportJob"
             :tool="exportJob?.tool || nativeToolHint"
@@ -344,13 +475,32 @@ const onRestoreCancelled = () => {
             :status-message="exportStatusMessage || ''"
             :progress="exportProgress"
             :is-running="isExporting"
+            :action-label="
+              lastExport?.saveDirectoryPath && !isExporting ? 'Open' : undefined
+            "
+            @action="openExportSaveDirectory"
           />
 
           <div class="grid gap-4 xl:grid-cols-[minmax(0,1fr)_280px]">
-            <div class="rounded-lg border p-4">
+            <div class="rounded-lg border p-4 space-y-4">
+              <NativeToolSelector
+                v-if="desktopRuntimeSelectionEnabled"
+                v-model="exportToolSelection"
+                title="Backup Driver"
+                description="Choose the exact executable OrcaQ should run for this backup job."
+                :detected-paths="exportToolPaths"
+                :tool-choices="exportToolChoices"
+                :disabled="isExporting"
+              />
+
               <ExportOptionsForm
                 :schemas="availableSchemas"
+                :schema-loading="isSchemaLoading"
+                :schema-error="schemaLoadError"
                 :loading="isExporting"
+                :disabled="
+                  desktopRuntimeSelectionEnabled && !canExportWithRuntime
+                "
                 :format-options="nativeBackupFormats"
                 :default-format="capability?.defaultExportFormat"
                 :tool-hint="nativeToolHint"
@@ -426,12 +576,25 @@ const onRestoreCancelled = () => {
                 @clear="clearSelectedFile"
               />
 
+              <NativeToolSelector
+                v-if="desktopRuntimeSelectionEnabled"
+                v-model="importToolSelection"
+                title="Restore Driver"
+                description="Choose the executable HeraQ should use to restore this backup artifact."
+                :detected-paths="importToolPaths"
+                :tool-choices="importToolChoices"
+                :disabled="isImporting"
+              />
+
               <ImportOptionsForm v-model="importOptions" />
 
               <Button
                 class="w-full"
-                size="lg"
-                :disabled="!selectedFile || isImporting"
+                :disabled="
+                  !selectedFile ||
+                  isImporting ||
+                  (desktopRuntimeSelectionEnabled && !canImportWithRuntime)
+                "
                 @click="onRestoreClick"
               >
                 <Icon
