+++ components/modules/database-tools/backup-restore/hooks/useDatabaseExport.ts
@@ -1,15 +1,47 @@
 import { getConnectionParams } from '@/core/helpers/connection-helper';
 import { useStreamingDownload } from '~/core/composables/useStreamingDownload';
 import { getNativeBackupExtension } from '~/core/constants/database-backup';
+import { formatBytes } from '~/core/helpers';
 import { type Connection } from '~/core/stores';
 import type {
   ExportDatabaseRequest,
   ExportFormat,
   ExportOptions,
   NativeBackupRuntimeCapability,
+  NativeBackupRuntimeSelection,
   StartDatabaseTransferResponse,
 } from '~/core/types';
 import { useDatabaseTransferJob } from './useDatabaseTransferJob';
+
+interface ExportExecutionRequest {
+  options: ExportOptions;
+  runtime?: NativeBackupRuntimeSelection;
+  promptForSaveLocation?: boolean;
+  saveDirectoryPath?: string;
+  saveFilePath?: string;
+}
+
+const sanitizeBackupFileSegment = (value: string) =>
+  (value || 'database')
+    .trim()
+    .replace(/[\\/:*?"<>|]+/g, '-')
+    .replace(/\s+/g, '-')
+    .replace(/-+/g, '-')
+    .replace(/^-|-$/g, '') || 'database';
+
+const createBackupTimestamp = () =>
+  new Date().toISOString().replace(/[:.]/g, '-');
+
+const joinDirectoryAndFileName = (directoryPath: string, fileName: string) =>
+  `${directoryPath.replace(/[\\/]$/, '')}/${fileName}`;
+
+const formatDuration = (durationMs: number) => {
+  if (!Number.isFinite(durationMs) || durationMs <= 0) {
+    return '0.0s';
+  }
+
+  return `${(durationMs / 1000).toFixed(1)}s`;
+};
 
 /**
  * Composable for database export operations
@@ -19,7 +51,13 @@ export const useDatabaseExport = (
   capability?: Ref<NativeBackupRuntimeCapability | null>
 ) => {
   const error = ref<string | null>(null);
-  const lastExport = ref<{ fileName: string; duration: number } | null>(null);
+  const lastExport = ref<{
+    fileName: string;
+    duration: number;
+    size: number;
+    saveDirectoryPath?: string;
+    saveFilePath?: string;
+  } | null>(null);
   const transferJob = useDatabaseTransferJob();
   const { downloadStream, isDownloading, downloadedBytes } =
     useStreamingDownload();
@@ -28,6 +66,10 @@ export const useDatabaseExport = (
     () => transferJob.isRunning.value || isDownloading.value
   );
   const progress = computed(() => {
+    if (lastExport.value) {
+      return 100;
+    }
+
     if (isDownloading.value) {
       return 100;
     }
@@ -35,8 +77,12 @@ export const useDatabaseExport = (
     return transferJob.progress.value;
   });
   const statusMessage = computed(() => {
+    if (lastExport.value) {
+      return `Backup is ready. ${formatBytes(lastExport.value.size)} generated in ${formatDuration(lastExport.value.duration)}.`;
+    }
+
     if (isDownloading.value) {
-      return `Downloading backup artifact (${downloadedBytes.value} bytes)...`;
+      return `Saving backup artifact (${formatBytes(downloadedBytes.value)})...`;
     }
 
     return transferJob.message.value;
@@ -48,14 +94,20 @@ export const useDatabaseExport = (
    */
   const exportDatabase = async (
     databaseName: string,
-    options: ExportOptions
+    request: ExportExecutionRequest
   ): Promise<boolean> => {
     if (!connection.value) {
       error.value = 'No database connection provided';
       return false;
     }
 
-    if (capability?.value && !capability.value.exportAvailable) {
+    const { options, runtime, promptForSaveLocation } = request;
+
+    if (
+      capability?.value &&
+      !capability.value.exportAvailable &&
+      !runtime?.executablePath?.trim()
+    ) {
       error.value = capability.value.exportMessage;
       return false;
     }
@@ -64,6 +116,30 @@ export const useDatabaseExport = (
     lastExport.value = null;
 
     try {
+      let saveDirectoryPath = request.saveDirectoryPath?.trim() || undefined;
+      let saveFilePath = request.saveFilePath?.trim() || undefined;
+      const fileName = `${sanitizeBackupFileSegment(databaseName)}_backup_${createBackupTimestamp()}${getExtension(connection.value?.type, options.format)}`;
+
+      if (
+        !saveDirectoryPath &&
+        !saveFilePath &&
+        promptForSaveLocation &&
+        window.electronAPI?.window.pickDirectory
+      ) {
+        const pickedPath = await window.electronAPI.window.pickDirectory();
+
+        if (!pickedPath) {
+          error.value = 'No save location was selected.';
+          return false;
+        }
+
+        saveDirectoryPath = pickedPath;
+      }
+
+      if (!saveFilePath && saveDirectoryPath) {
+        saveFilePath = joinDirectoryAndFileName(saveDirectoryPath, fileName);
+      }
+
       const response = await $fetch<StartDatabaseTransferResponse>(
         '/api/database-export/export-database',
         {
@@ -72,6 +148,7 @@ export const useDatabaseExport = (
             ...getConnectionParams(connection.value),
             databaseName,
             options,
+            runtime,
           } as ExportDatabaseRequest & Record<string, unknown>,
         }
       );
@@ -91,9 +168,14 @@ export const useDatabaseExport = (
       const downloadResult = await downloadStream({
         url: snapshot.downloadUrl,
         method: 'GET',
-        filename:
-          snapshot.downloadFileName ||
-          `${databaseName}_backup${getExtension(connection.value?.type, options.format)}`,
+        filename: saveFilePath
+          ? fileName
+          : snapshot.downloadFileName || fileName,
+        saveFilePath,
+        openPath: saveDirectoryPath,
+        successTitle: 'Backup is ready',
+        getSuccessDescription: sizeBytes =>
+          `${formatBytes(sizeBytes)} generated in ${formatDuration(snapshot.duration || 0)}.`,
       });
 
       if (!downloadResult.success) {
@@ -102,8 +184,11 @@ export const useDatabaseExport = (
       }
 
       lastExport.value = {
-        fileName: snapshot.downloadFileName,
+        fileName: saveFilePath ? fileName : snapshot.downloadFileName,
         duration: snapshot.duration || 0,
+        size: downloadResult.size || 0,
+        saveDirectoryPath,
+        saveFilePath,
       };
 
       return true;
