+++ components/modules/database-tools/backup-restore/types/backup-restore.types.ts
@@ -1,3 +1,5 @@
+import type { NativeBackupToolName } from '~/core/types';
+
 export type BackupRestoreTab = 'export' | 'import';
 
 export interface BackupRestoreSection {
@@ -9,3 +11,10 @@ export interface BackupRestoreSection {
 
 export type ToolTab = BackupRestoreTab;
 export type ToolSection = BackupRestoreSection;
+
+export interface BackupRestoreRuntimeToolSelection {
+  useCustomPath: boolean;
+  selectedPath: string;
+  customPath: string;
+  customTool: NativeBackupToolName | '';
+}
