+++ core/data/changelogs/1.1.7.md
@@ -1,0 +1,19 @@
+---
+version: '1.1.7'
+date: '2026-05-26'
+---
+
+# 🎉 v1.1.7 Changelog
+
+## 🚀 Key Features & Improvements
+
+- **Fullscreen Result Views**: Added a zoom/fullscreen button to the raw query result tabs. Users can now expand query result grids to full-screen mode to analyze large datasets, and exit easily by pressing `Escape`.
+- **Auto-Download Updates**: Added support to download application updates automatically in the background (configurable via a new toggle in Desktop Settings).
+- **Native Tool Path Injection (macOS)**: Automatically injects typical terminal `$PATH` directories (Homebrew, `/usr/local/bin`, and `Postgres.app`) for packaged macOS apps to guarantee native backup tools (`pg_dump`, `mysqldump`, `sqlite3`) execute reliably.
+- **Improved Backup & Restore UX**:
+  - Dynamically prompts for a save folder when backing up instead of requiring pre-selection.
+  - Aligned custom path input heights with select dropdowns and removed uppercase styling from labels.
+  - Added a **Check Again** scan button to immediately verify newly installed CLI backup tools.
+- **Database Connection Enhancements**:
+  - Added arrow key and `Enter` keyboard navigation when choosing a database type.
+  - Wrapped connection credentials in standard HTML forms to allow submission by pressing `Enter`.
