+++ electron/ipc/window.ts
@@ -6,7 +6,9 @@ import {
   dialog,
   type BrowserWindow,
   type OpenDialogOptions,
+  type SaveDialogOptions,
 } from 'electron';
+import { writeFile } from 'node:fs/promises';
 import path from 'node:path';
 import { getElectronLogPath } from '../logger';
 import { clearPersistedUserData } from '../persist/store';
@@ -76,6 +78,63 @@ export function registerWindowHandlers(
     }
 
     return result.filePaths[0] || null;
+  });
+
+  ipcMain.handle(
+    'window:pick-save-file',
+    async (_event, options?: SaveDialogOptions) => {
+      const mainWindow = getMainWindow();
+      const result = mainWindow
+        ? await dialog.showSaveDialog(mainWindow, options || {})
+        : await dialog.showSaveDialog(options || {});
+
+      if (result.canceled) {
+        return null;
+      }
+
+      return result.filePath || null;
+    }
+  );
+
+  ipcMain.handle('window:pick-directory', async () => {
+    const mainWindow = getMainWindow();
+    const options: OpenDialogOptions = {
+      title: 'Choose Backup Folder',
+      properties: ['openDirectory', 'createDirectory'],
+    };
+    const result = mainWindow
+      ? await dialog.showOpenDialog(mainWindow, options)
+      : await dialog.showOpenDialog(options);
+
+    if (result.canceled) {
+      return null;
+    }
+
+    return result.filePaths[0] || null;
+  });
+
+  ipcMain.handle(
+    'window:write-file',
+    async (_event, payload?: { filePath?: string; data?: Uint8Array }) => {
+      if (!payload?.filePath) {
+        throw new Error('A file path is required to save the downloaded file.');
+      }
+
+      if (!payload.data) {
+        throw new Error('Downloaded file data is missing.');
+      }
+
+      await writeFile(payload.filePath, Buffer.from(payload.data));
+      return payload.filePath;
+    }
+  );
+
+  ipcMain.handle('window:open-path', async (_event, targetPath?: string) => {
+    if (!targetPath) {
+      throw new Error('A path is required to open a folder.');
+    }
+
+    return await shell.openPath(targetPath);
   });
 
   ipcMain.handle('window:get-storage-path', () => {
