+++ npx-package/package.json
@@ -1,6 +1,6 @@
 {
   "name": "orcaq",
-  "version": "1.1.6",
+  "version": "1.1.7",
   "description": "The open source | Next Generation database editor - run locally via npx",
   "author": "Cinny",
   "license": "MIT",
