+++ package.json
@@ -9,7 +9,7 @@
     "url": "https://github.com/cin12211"
   },
   "license": "MIT",
-  "version": "1.1.6",
+  "version": "1.1.7",
   "main": "out/electron/main.js",
   "scripts": {
     "dev": "npm run nuxt:dev",
