+++ server/infrastructure/database/backup/logical/index.ts
@@ -1,0 +1,6 @@
+export type {
+  BackupConnectionParams,
+  LogicalBackupTable,
+  LogicalDatabaseBackup,
+} from './types';
+export { parseLogicalBackup } from './shared';
