+++ server/infrastructure/database/backup/native/index.ts
@@ -1,0 +1,11 @@
+import { applyPlatformPathFallbacks } from './constants';
+
+applyPlatformPathFallbacks();
+
+export type { NativeBackupToolName } from '~/core/types';
+export type {
+  NativeBackupCapability,
+  NativeBackupFormatOption,
+} from './constants';
+export * from './capability';
+export * from './runtime';
